Add a .travis file and update tests for it
