+++ .travis.yaml
@@ -1,0 +1,6 @@
+language: go
+go:
+  - 1.14.6
+script:
+  - |
+    make unit-test
