+++ Makefile
@@ -1,2 +1,5 @@
 generate:
 	go generate ./...
+
+unit-test:
+	go test ./...
