+++ azure/azure_kv.go
@@ -6,9 +6,8 @@ import (
 	"errors"
 	"time"
 
-	"github.com/Azure/go-autorest/autorest/to"
-
 	"github.com/Azure/azure-sdk-for-go/services/keyvault/2016-10-01/keyvault"
+	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/libopenstorage/secrets"
 	"github.com/portworx/sched-ops/task"
 )
