+++ dcos/dcos_integration_test.go
@@ -1,3 +1,5 @@
+// +build integration
+
 package dcos
 
 import (
