+++ vault/vault_integration_test.go
@@ -1,4 +1,4 @@
-// +build e2e
+// +build integration
 
 package vault
